Refactor: Move sd2_clip.py to text_encoders folder.
