+++ comfy/text_encoders/sd2_clip.py
