+++ comfy/text_encoders/sd2_clip_config.json
